fixed rng hopefully, modifying everything to use unsigned and get_rand()

+++ warp_header.h
@@ -1,7 +1,7 @@
 // host calls
 void print_banner();
 // device calls
-void set_positions_rand( unsigned, unsigned, unsigned, source_point * , float *  , float  * );
+void set_positions_rand( unsigned, unsigned, unsigned, source_point * , unsigned *  , float  * );
 void copy_points(unsigned , unsigned , unsigned , unsigned*  , unsigned  , unsigned *  , source_point *  , source_point * , float*, float*);
 void sample_fission_spectra(unsigned , unsigned , unsigned , unsigned* , unsigned* , unsigned* , unsigned* , float * , float *, source_point* , float** );
 void sample_fixed_source( unsigned,unsigned,unsigned,unsigned*,float*,float*,source_point*);
@@ -10,7 +10,7 @@ void microscopic( unsigned ,  unsigned , unsigned , unsigned , unsigned*, unsign
 void tally_spec( unsigned ,  unsigned , unsigned , unsigned*, source_point * , float* , float * , unsigned * , unsigned*, unsigned*);
 void escatter( cudaStream_t, unsigned , unsigned, unsigned , unsigned*, unsigned* , unsigned* , float* , float*, source_point* , unsigned*, float*, unsigned*, float**);
 void iscatter( cudaStream_t, unsigned , unsigned , unsigned , unsigned*, unsigned* , unsigned * , float * , float *, source_point *  ,unsigned * , float* , float* , unsigned* , float**, float**);
-void cscatter( cudaStream_t, unsigned , unsigned , unsigned , unsigned*, unsigned* , unsigned * , float * , float *, source_point *  ,unsigned * , float* , float* , unsigned* , float**, float**);
+void cscatter( cudaStream_t, unsigned , unsigned , unsigned , unsigned*, unsigned* , unsigned * , unsigned * , float *, source_point *  ,unsigned * , float* , float* , unsigned* , float**, float**);
 void fission(  cudaStream_t, unsigned , unsigned, unsigned, unsigned , unsigned*, unsigned* , unsigned* , unsigned*  , float * , unsigned* , float**);
 void absorb(   cudaStream_t, unsigned , unsigned , unsigned*, unsigned*  , unsigned* );
 void find_E_grid_index( unsigned , unsigned , unsigned , unsigned*,float * , float* , unsigned *, unsigned* );
@@ -18,6 +18,6 @@ void find_E_grid_index_quad( unsigned, unsigned,  unsigned,  unsigned, unsigned*
 void make_mask( unsigned, unsigned, unsigned*, unsigned*, unsigned, unsigned);
 void print_histories(unsigned, unsigned, unsigned *, unsigned*, source_point*, float*, unsigned*,unsigned*);
 void pop_secondaries( unsigned, unsigned, unsigned, unsigned* , unsigned* , unsigned* , unsigned* , unsigned*, unsigned*, source_point* , float* , float* , float** );
-void pop_source( unsigned, unsigned, unsigned, unsigned*, unsigned* , unsigned* , unsigned* , unsigned* , unsigned*, unsigned*, source_point* , float* , float* , float** , source_point*, float*, float*);
-void rebase_yield( unsigned , unsigned , unsigned , float , float* , unsigned* );
+void pop_source( unsigned, unsigned, unsigned, unsigned*, unsigned* , unsigned* , unsigned* , unsigned* , unsigned*, unsigned*, source_point* , float* , unsigned* , float** , source_point*, float*, float*);
+void rebase_yield( unsigned , unsigned , unsigned , float , unsigned* , unsigned* );
 void flip_done(unsigned , unsigned , unsigned* );
